Modernize the package
Use uv.
Drop support for unsupported Pythons.
Use GitHub actions for build and deploy.
Use pyproject.toml instead of setup.cfg.
Move coverage config to pyproject.toml.
