--- .travis.yml
@@ -1,35 +1,0 @@
-dist: trusty
-sudo: false
-language: python
-python:
-  - 2.7
-  - 3.4
-  - 3.5
-  - 3.6
-install:
- - pip install tox-travis wheel codecov coverage
- - python setup.py bdist_wheel
-script: tox --installpkg ./dist/*.whl
-after_success: coverage combine && codecov
-stages:
-  - test
-  - name: deploy
-    if: tag IS present
-jobs:
-  fast_finish: true
-  include:
-    - python: 3.6
-      env: TOXENV=desc
-    - stage: deploy
-      python: 3.6
-      install: true
-      script: true
-      after_success: true
-      deploy:
-        provider: pypi
-        user: ryanhiebert-auto
-        password:
-          secure: d5Z3xlI2kdy+GzR0Zj805NSM3YHc0j3D2+zClQ/zLfvAoWzHiOIXkWL0xVfSKVcbzI0Jx05FiHQt3FoWjnYmYqZVFcyXM3ehiyOsw/k2VIw0wu7+qyJIHeGrkr0GmXd8oqHdHNmBPXubI/I9kDgcEdALDboxxQcK1phrOtjUgNiBlIeKgEoJw4qo8CHL9lbyfVkXzoBV7wxWwp4nE+SkZcvex7i+i1zAVp5Akrr3oEYu2zrFpiYN2+fccNQLTgoSS1YRiWGWPbYd2+d2dZOf+pNnlovQWLZF+Urw47YdrE3RN4qsm4i8O/BzVriVvvShLONDLHBkeFcFYfB+El9n1uc0UwvwslJeyVEdu1QB5sv/Uq0Ge2vVZEyZoW2qzfvGCFj8yNT+tpchagtuiyklwHOi+TuKjCYUDVLgGB6vJmaSTKnpGKlCOUTUnbKD/zEAUkk7XMPa9MtrF0AYY+VJ0oBQDftSH89jNtC0Bv9mh9HBx6oDRV+icvjK4dn3BxCTFwLsrHR7dyZ+SteK5EM25cFsX+3voZoMePIL78+8XsAKNhG0GWWQGaV6lij+XI9Xtx8cy2e/uNZwn5R9UoNprfZ61cPbc5Gk9VnZSrlqEY9PiHWHrhZuiC30isPW4pQciB4kksRb+Kbx2KnoZfwQbgAIbBARe510U47RN8zJdNM=
-        on:
-          tags: true
-        distributions: bdist_wheel
