+++ pyproject.toml
@@ -1,0 +1,44 @@
+[project]
+name = "django-cmd"
+version = "2.0"
+description = "Get a django command"
+authors = [{ name = "Ryan Hiebert", email = "ryan@ryanhiebert.com" }]
+license = "MIT"
+readme = "README.rst"
+dependencies = ["Django"]
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Framework :: Django :: 1.8",
+    "Framework :: Django :: 1.10",
+    "Framework :: Django :: 1.11",
+    "Framework :: Django :: 2.0",
+]
+
+[project.urls]
+homepage = "https://github.com/ryanhiebert/django-cmd"
+repository = "https://github.com/ryanhiebert/django-cmd"
+
+[project.scripts]
+django = "django_cmd:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[dependency-groups]
+dev = ["pytest", "pytest-mock", "coverage_pth", "ruff", "isort"]
+
+[tool.coverage.run]
+branch = true
+parallel = true
+source = "django_cmd"
