+++ tox.ini
@@ -1,30 +1,31 @@
 [tox]
+isolated_build = True
 envlist =
-    py{27,34,35}-dj18
-    py{27,34,35}-dj110
-    py{27,34,35,36}-dj111
-    py{34,35,36}-dj20
-    desc
+    dj22-py{38,39                }
+    dj30-py{38,39                }
+    dj31-py{38,39                }
+    dj32-py{38,39,310            }
+    dj40-py{38,39,310            }
+    dj41-py{38,39,310,311        }
+    dj42-py{   39,310,311        }
+    dj50-py{      310,311,312    }
+    dj51-py{      310,311,312,313}
 
 [testenv]
-deps =
-    pytest
-    pytest-mock
-    coverage_pth
-    dj18: Django>=1.8,<1.9
-    dj110: Django>=1.10,<1.11
-    dj111: Django>=1.11,<1.12
-    dj20: Django>=2.0,<2.1
 setenv =
     COVERAGE_PROCESS_START=.coveragerc
-commands = {posargs:pytest}
-
-[testenv:desc]
+dependency_groups = dev
 deps =
-    docutils
-    Pygments
-commands =
-    python setup.py check --restructuredtext --strict
+    dj22: Django~=2.2.0
+    dj30: Django~=3.0.0
+    dj31: Django~=3.1.0
+    dj32: Django~=3.2.0
+    dj40: Django~=4.0.0
+    dj41: Django~=4.1.0
+    dj42: Django~=4.2.0
+    dj50: Django~=5.0.0
+    dj51: Django~=5.1.0
+commands = {posargs:pytest}
 
 [coverage:run]
 branch = True
